requirements.txt
Flask
Werkzeug
RPi.GPIO